tested mouse movement behaviors of Cursor.as
I have introduced a class called FakeSprite that allows me to manipulate the read only property mouseX and mouseY for test purposes on DisplayObjects. Main subclasses FakeSprite for the moment because it gets down cast from DisplayObjectContainer to Main in Cursor, and i need to control the container's mouseX and mouseY values for tests of Cursor. I plan more work on CursorTest.as for tomorrow, as I still have to test clicking behaviors on Cursor.'